ground texture, unit movement rotation, construction upgrade

+++ Project -v1.0.2 - 4.2.0/Project -v1.0.2 - 4.2.0/Assets/CostBox.cs
@@ -72,7 +72,7 @@ public class CostBox : MonoBehaviour {
 
 
 			if (input.myCost.ResourceOne > 0) {
-				resOne.text = "" + input.myCost.ResourceOne;
+				resOne.text = "" + (int)input.myCost.ResourceOne;
 				OreIcon.enabled =true;
 				if (order.reasonList.Contains (continueOrder.reason.resourceOne)) {
 					resOne.color = Color.red;
@@ -128,7 +128,7 @@ public class CostBox : MonoBehaviour {
 					resTwo.color = teal;
 				}
 
-				resTwo.text = "Gas: " + input.myCost.ResourceTwo;
+				resTwo.text = "Gas: " + (int)input.myCost.ResourceTwo;
 			} else {
 				resTwo.text = "";
 			}
